fastapi_stepik/fastapi_ecommerce/requirements.txt
aiosqlite==0.20.0
alembic==1.13.1
annotated-types==0.6.0
anyio==4.2.0
async-timeout==4.0.3
asyncpg==0.29.0
bcrypt==4.0.1
cffi==1.16.0
click==8.1.7
cryptography==42.0.5
ecdsa==0.18.0
exceptiongroup==1.2.2
fastapi==0.109.2
greenlet==3.0.3
h11==0.14.0
httptools==0.6.1
idna==3.6
loguru==0.7.2
Mako==1.3.2
MarkupSafe==2.1.5
passlib==1.7.4
pyasn1==0.6.0
pycparser==2.22
pydantic==2.6.1
pydantic_core==2.16.2
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.9
python-slugify==8.0.4
PyYAML==6.0.1
rsa==4.9
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.27
starlette==0.36.3
text-unidecode==1.3
typing_extensions==4.9.0
uvicorn==0.27.1
uvloop==0.19.0
watchfiles==0.21.0
websockets==12.0
